requirements.txt
numpy==1.16.4
pandas==0.24.2
Pillow==5.2.0
scikit-learn==0.21.2
streamlit==0.47.4
tensorflow==1.14.0
tqdm==4.36.1
